refactor(backend): DealEventListener to package deal
