+++ backend/src/main/java/com/onlyoffice/docspacepipedrive/events/deal/AddRoomToPipedriveDealEvent.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.onlyoffice.docspacepipedrive.events;
+package com.onlyoffice.docspacepipedrive.events.deal;
 
 import com.onlyoffice.docspacepipedrive.client.pipedrive.dto.PipedriveDeal;
 import lombok.Getter;
@@ -24,11 +24,13 @@ import org.springframework.context.ApplicationEvent;
 
 
 @Getter
-public class AddVisibleEveryoneForPipedriveDealEvent extends ApplicationEvent {
+public class AddRoomToPipedriveDealEvent extends ApplicationEvent {
     private final PipedriveDeal pipedriveDeal;
+    private final Long roomId;
 
-    public AddVisibleEveryoneForPipedriveDealEvent(final Object source, final PipedriveDeal pipedriveDeal) {
+    public AddRoomToPipedriveDealEvent(final Object source, final PipedriveDeal pipedriveDeal, final Long roomId) {
         super(source);
         this.pipedriveDeal = pipedriveDeal;
+        this.roomId = roomId;
     }
 }
