+++ backend/src/main/java/com/onlyoffice/docspacepipedrive/events/deal/RemoveVisibleEveryoneForPipedriveDealEvent.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.onlyoffice.docspacepipedrive.events;
+package com.onlyoffice.docspacepipedrive.events.deal;
 
 import com.onlyoffice.docspacepipedrive.client.pipedrive.dto.PipedriveDeal;
 import lombok.Getter;
