+++ backend/src/main/java/com/onlyoffice/docspacepipedrive/web/controller/RoomController.java
@@ -24,7 +24,7 @@ import com.onlyoffice.docspacepipedrive.client.pipedrive.PipedriveClient;
 import com.onlyoffice.docspacepipedrive.client.pipedrive.dto.PipedriveDeal;
 import com.onlyoffice.docspacepipedrive.entity.Client;
 import com.onlyoffice.docspacepipedrive.entity.Room;
-import com.onlyoffice.docspacepipedrive.events.AddRoomToPipedriveDealEvent;
+import com.onlyoffice.docspacepipedrive.events.deal.AddRoomToPipedriveDealEvent;
 import com.onlyoffice.docspacepipedrive.service.RoomService;
 import com.onlyoffice.docspacepipedrive.web.dto.room.RoomResponse;
 import com.onlyoffice.docspacepipedrive.web.mapper.RoomMapper;
