+++ backend/src/main/java/com/onlyoffice/docspacepipedrive/web/controller/WebhookController.java
@@ -24,10 +24,10 @@ import com.onlyoffice.docspacepipedrive.client.pipedrive.dto.PipedriveUser;
 import com.onlyoffice.docspacepipedrive.client.pipedrive.dto.PipedriveUserSettings;
 import com.onlyoffice.docspacepipedrive.entity.Client;
 import com.onlyoffice.docspacepipedrive.entity.User;
-import com.onlyoffice.docspacepipedrive.events.AddFollowersToPipedriveDealEvent;
-import com.onlyoffice.docspacepipedrive.events.AddVisibleEveryoneForPipedriveDealEvent;
-import com.onlyoffice.docspacepipedrive.events.RemoveFollowersFromPipedriveDealEvent;
-import com.onlyoffice.docspacepipedrive.events.RemoveVisibleEveryoneForPipedriveDealEvent;
+import com.onlyoffice.docspacepipedrive.events.deal.AddFollowersToPipedriveDealEvent;
+import com.onlyoffice.docspacepipedrive.events.deal.AddVisibleEveryoneForPipedriveDealEvent;
+import com.onlyoffice.docspacepipedrive.events.deal.RemoveFollowersFromPipedriveDealEvent;
+import com.onlyoffice.docspacepipedrive.events.deal.RemoveVisibleEveryoneForPipedriveDealEvent;
 import com.onlyoffice.docspacepipedrive.exceptions.PipedriveAccessDeniedException;
 import com.onlyoffice.docspacepipedrive.exceptions.RoomNotFoundException;
 import com.onlyoffice.docspacepipedrive.manager.PipedriveActionManager;
